JDK-8380447: Fix linter warnings in test/hotspot/jtreg/compiler/lib
