+++ test/hotspot/jtreg/compiler/lib/generators/EmptyGeneratorException.java
@@ -28,6 +28,7 @@ package compiler.lib.generators;
  * set of values. For example, bounds such as [1, 0] cause an EmptyGeneratorException. Another example would be
  * restricting a uniform integer generator over the range [0, 1] to [10, 11].
  */
+@SuppressWarnings("serial")
 public class EmptyGeneratorException extends RuntimeException {
     public EmptyGeneratorException() {}
 }
