+++ test/hotspot/jtreg/compiler/lib/generators/Generators.java
@@ -258,6 +258,7 @@ public final class Generators {
      * An overload for restrictable generators exists.
      */
     @SafeVarargs
+    @SuppressWarnings("varargs")
     public final <T> Generator<T> mixed(List<Integer> weights, Generator<T>... generators) {
         return new MixedGenerator<>(this, Arrays.asList(generators), weights);
     }
@@ -279,6 +280,7 @@ public final class Generators {
      * generator.
      */
     @SafeVarargs
+    @SuppressWarnings("varargs")
     public final <T extends Comparable<T>> RestrictableGenerator<T> mixed(List<Integer> weights, RestrictableGenerator<T>... generators) {
         return new RestrictableMixedGenerator<>(this, Arrays.asList(generators), weights);
     }
