+++ test/hotspot/jtreg/compiler/lib/ir_framework/driver/SuccessOnlyConstraintException.java
@@ -28,6 +28,7 @@ import compiler.lib.ir_framework.driver.irmatching.irrule.constraint.Constraint;
 /**
  * Exception used to signal that a {@link Constraint} should always succeed.
  */
+@SuppressWarnings("serial")
 public class SuccessOnlyConstraintException extends RuntimeException {
     public SuccessOnlyConstraintException(String message) {
         super("Unhandled SuccessOnlyConstraintException, should have created a Constraint that always succeeds:" + System.lineSeparator() + message);
