+++ test/hotspot/jtreg/compiler/lib/ir_framework/shared/TestFormatException.java
@@ -26,6 +26,7 @@ package compiler.lib.ir_framework.shared;
 /**
  * Exception that is thrown if a JTreg test violates the supported format by the test framework.
  */
+@SuppressWarnings("serial")
 public class TestFormatException extends RuntimeException {
     public TestFormatException(String message) {
         super(message);
