+++ test/hotspot/jtreg/compiler/lib/ir_framework/test/SetupArgumentsProvider.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2026, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package compiler.lib.ir_framework.test;
+
+import compiler.lib.ir_framework.SetupInfo;
+import compiler.lib.ir_framework.shared.TestRunException;
+
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+
+/**
+ * Used for @Arguments(setup = "setupMethodName") to specify a setup method to provide arguments
+ * and possibly set fields.
+ */
+final class SetupArgumentsProvider implements ArgumentsProvider {
+    Method setupMethod;
+
+    SetupArgumentsProvider(Method setupMethod) {
+        this.setupMethod = setupMethod;
+    }
+
+    @Override
+    public Object[] getArguments(Object invocationTarget, int invocationCounter) {
+        Object target = Modifier.isStatic(setupMethod.getModifiers()) ? null
+                : invocationTarget;
+        try {
+            if (setupMethod.getParameterCount() == 1) {
+                return (Object[]) setupMethod.invoke(target, new SetupInfo(invocationCounter));
+            } else {
+                return (Object[]) setupMethod.invoke(target);
+            }
+        } catch (Exception e) {
+            throw new TestRunException("There was an error while invoking setup method " +
+                    setupMethod + " on " + target, e);
+        }
+    }
+}
