+++ test/hotspot/jtreg/compiler/lib/template_framework/DataName.java
@@ -202,7 +202,7 @@ public record DataName(String name, DataName.Type type, boolean mutable, int wei
          *                                       {@link #subtypeOf}, {@link #supertypeOf} or {@link #exactOf}.
          */
         public Token sample(Function<DataName, ScopeToken> function) {
-            return new NameSampleToken<>(predicate(), null, null, function);
+            return new NameSampleToken<>(DataName.class, predicate(), null, null, function);
         }
 
         /**
@@ -244,7 +244,7 @@ public record DataName(String name, DataName.Type type, boolean mutable, int wei
          *                                       {@link #subtypeOf}, {@link #supertypeOf} or {@link #exactOf}.
          */
         public Token sampleAndLetAs(String name, String type) {
-            return new NameSampleToken<DataName>(predicate(), name, type, n -> Template.transparentScope());
+            return new NameSampleToken<>(DataName.class, predicate(), name, type, n -> Template.transparentScope());
         }
 
         /**
@@ -284,7 +284,7 @@ public record DataName(String name, DataName.Type type, boolean mutable, int wei
          *                                       {@link #subtypeOf}, {@link #supertypeOf} or {@link #exactOf}.
          */
         public Token sampleAndLetAs(String name) {
-            return new NameSampleToken<DataName>(predicate(), name, null, n -> Template.transparentScope());
+            return new NameSampleToken<>(DataName.class, predicate(), name, null, n -> Template.transparentScope());
         }
 
         /**
@@ -326,7 +326,7 @@ public record DataName(String name, DataName.Type type, boolean mutable, int wei
          *                                       {@link #subtypeOf}, {@link #supertypeOf} or {@link #exactOf}.
          */
         public Token toList(Function<List<DataName>, ScopeToken> function) {
-            return new NamesToListToken<>(predicate(), function);
+            return new NamesToListToken<>(DataName.class, predicate(), function);
         }
 
         /**
@@ -340,7 +340,7 @@ public record DataName(String name, DataName.Type type, boolean mutable, int wei
          *                                       {@link #subtypeOf}, {@link #supertypeOf} or {@link #exactOf}.
          */
         public Token forEach(Function<DataName, ScopeToken> function) {
-            return new NameForEachToken<>(predicate(), null, null, function);
+            return new NameForEachToken<>(DataName.class, predicate(), null, null, function);
         }
 
         /**
@@ -364,7 +364,7 @@ public record DataName(String name, DataName.Type type, boolean mutable, int wei
          *                                       {@link #subtypeOf}, {@link #supertypeOf} or {@link #exactOf}.
          */
         public Token forEach(String name, String type, Function<DataName, ScopeToken> function) {
-            return new NameForEachToken<>(predicate(), name, type, function);
+            return new NameForEachToken<>(DataName.class, predicate(), name, type, function);
         }
     }
 }
